api/requirements.txt
anyio==3.7.1
bcrypt==4.1.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.13
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
websockets==12.0
wheel==0.41.2
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9